requirements.txt
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.2
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
pyyaml>=6.0
tqdm>=4.65.0
bert-score>=0.3.13
openai>=1.3.0
anthropic>=0.18.0
google-generativeai>=0.3.0
pandas>=2.0.0
matplotlib>=3.7.0
jsonlines>=3.1.0
pytest>=7.4.0
